+++ example/gitlab_dart_lib_example.dart
@@ -1,6 +1,0 @@
-import 'package:gitlab_dart_lib/gitlab_dart_lib.dart';
-
-main() {
-  var awesome = new Awesome();
-  print('awesome: ${awesome.isAwesome}');
-}